+++ argos/widgets/albumswindow.py
@@ -1,7 +1,6 @@
 import logging
 import re
 import threading
-import time
 from enum import IntEnum
 from pathlib import Path
 from typing import List, Optional
@@ -11,7 +10,9 @@ from gi.repository.GdkPixbuf import Pixbuf
 
 from argos.message import MessageType
 from argos.utils import elide_maybe
+from argos.widgets.albumdetailsbox import AlbumDetailsBox
 from argos.widgets.albumsbrowsingprogressbox import AlbumsBrowsingProgressBox
+from argos.widgets.condensedplayingbox import CondensedPlayingBox
 from argos.widgets.utils import default_image_pixbuf, scale_album_image
 
 LOGGER = logging.getLogger(__name__)
@@ -28,13 +29,14 @@ class AlbumStoreColumns(IntEnum):
 
 
 @Gtk.Template(resource_path="/io/github/orontee/Argos/ui/albums_window.ui")
-class AlbumsWindow(Gtk.Overlay):
+class AlbumsWindow(Gtk.Box):
     __gtype_name__ = "AlbumsWindow"
 
-    __gsignals__ = {"album-selected": (GObject.SIGNAL_RUN_FIRST, None, (str,))}
-
+    albums_overlay: Gtk.Overlay = Gtk.Template.Child()
+    albums_stack: Gtk.Stack = Gtk.Template.Child()
     albums_view: Gtk.IconView = Gtk.Template.Child()
 
+    album_details_box = GObject.Property(type=AlbumDetailsBox)
     filtered_albums_store = GObject.Property(type=Gtk.TreeModelFilter)
     filtering_text = GObject.Property(type=str)
 
@@ -55,6 +57,9 @@ class AlbumsWindow(Gtk.Overlay):
             target_width=self.albums_image_size,
         )
 
+        self.props.album_details_box = AlbumDetailsBox(application)
+        self.albums_stack.add_named(self.props.album_details_box, "album_details_page")
+
         albums_store = Gtk.ListStore(str, str, str, str, Pixbuf, str, str)
         self.props.filtered_albums_store = albums_store.filter_new()
         self.props.filtered_albums_store.set_visible_func(self._filter_album_row, None)
@@ -65,11 +70,13 @@ class AlbumsWindow(Gtk.Overlay):
         self.albums_view.set_pixbuf_column(AlbumStoreColumns.PIXBUF)
         self.albums_view.set_item_width(self.albums_image_size)
 
+        self.add(CondensedPlayingBox(application))
+
         if application.props.disable_tooltips:
             self.albums_view.props.has_tooltip = False
 
         self._progress_box = AlbumsBrowsingProgressBox()
-        self.add_overlay(self._progress_box)
+        self.albums_overlay.add_overlay(self._progress_box)
         self._progress_box.show_all()
         self.show_all()
 
@@ -212,6 +219,12 @@ class AlbumsWindow(Gtk.Overlay):
                     LOGGER.debug(f"No image path for {uri}")
                 store_iter = store.iter_next(store_iter)
 
+    def is_albums_page_visible(self) -> bool:
+        return self.albums_stack.get_visible_child_name() == "albums_page"
+
+    def select_albums_page(self) -> None:
+        self.albums_stack.set_visible_child_name("albums_page")
+
     @Gtk.Template.Callback()
     def albums_view_item_activated_cb(
         self, icon_view: Gtk.IconView, path: Gtk.TreePath
@@ -223,7 +236,15 @@ class AlbumsWindow(Gtk.Overlay):
         filtered_store = icon_view.get_model()
         store_iter = filtered_store.get_iter(path)
         uri = filtered_store.get_value(store_iter, AlbumStoreColumns.URI)
-        self.emit("album-selected", uri)
+
+        LOGGER.debug(f"Album {uri!r} selected")
+        self._app.send_message(
+            MessageType.COMPLETE_ALBUM_DESCRIPTION, {"album_uri": uri}
+        )
+
+        self.props.album_details_box.set_property("uri", uri)
+        self.props.album_details_box.show_now()
+        self.albums_stack.set_visible_child_name("album_details_page")
 
     def _on_albums_image_size_changed(
         self,